Renamed pagination file for category subpages
